Real-Time News & Research Summary Hub/requirements.txt
streamlit
requests
tavily-python
arxiv
python-dotenv
python-docx
transformers